Fixing paths issues
