+++ lib/api/package.json
@@ -5,7 +5,7 @@
   "main": "dist/index.js",
   "types": "types/index.d.ts",
   "scripts": {
-    "build": "tslint --config ../../tslint.json --project tsconfig.json && tsc",
+    "build": "tslint --config ../../tslint.json --project tsconfig.json && ttsc",
     "test": "for d in test/*/; do mocha ${d}/setup.ts ${d}/**/*.spec.ts -r ts-node/register -r tsconfig-paths/register; done"
   },
   "keywords": [
@@ -44,12 +44,15 @@
     "@types/node": "^10.5.5",
     "@types/sinon": "^4.3.3",
     "@types/sinon-chrome": "^2.2.4",
+    "@zerollup/ts-transform-paths": "^1.6.5",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "mocha": "^5.1.1",
     "sinon": "^5.1.1",
     "sinon-chrome": "^2.3.2",
     "ts-node": "^6.1.0",
-    "tslint": "^5.11.0"
+    "tsconfig-paths": "^3.7.0",
+    "tslint": "^5.11.0",
+    "ttypescript": "^1.5.5"
   }
 }
