+++ lib/api/src/tabs/chrome/events.ts
