+++ lib/api/src/tabs/extensions/Tabs.ts
