+++ lib/api/src/tabs/extensions/events.ts
