--- lib/api/src/tabs/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@internal/tabs",
-  "scripts": {
-    "test": "mocha test/setup.ts test/**/*.spec.ts -r ts-node/register"
-  }
-}
