+++ lib/api/src/tabs/safari/events.ts
