+++ lib/api/test/management/spec/permissions/SafariPermissions.spec.ts
@@ -3,7 +3,7 @@ import { assert } from 'chai'
 import { Container } from '@exteranto/core'
 import { Browser } from '@exteranto/core'
 
-import { Permission, PermissionManager } from '../../../src'
+import { Permission, PermissionManager } from '../../../../src'
 
 export const tests = () => {
   describe('Safari', () => {
