+++ lib/core/package.json
@@ -5,7 +5,7 @@
   "main": "dist/index.js",
   "types": "types/index.d.ts",
   "scripts": {
-    "build": "tslint --config ../../tslint.json --project tsconfig.json && tsc",
+    "build": "tslint --config ../../tslint.json --project tsconfig.json && ttsc",
     "test": "for d in test/*/; do mocha ${d}/setup.ts ${d}/**/*.spec.ts -r ts-node/register -r tsconfig-paths/register; done"
   },
   "keywords": [
@@ -39,12 +39,14 @@
     "@types/chai-as-promised": "^7.1.0",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.12.18",
+    "@zerollup/ts-transform-paths": "^1.6.5",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "mocha": "^5.1.1",
     "sinon": "^7.2.2",
     "ts-node": "^6.2.0",
     "tsconfig-paths": "^3.7.0",
-    "tslint": "^5.11.0"
+    "tslint": "^5.11.0",
+    "ttypescript": "^1.5.5"
   }
 }
