+++ lib/exceptions/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@exteranto/exceptions",
-  "version": "1.1.2",
+  "version": "2.0.0",
   "description": "The Exceptions Package for the Exteranto Framework.",
   "main": "dist/index.js",
   "types": "types/index.d.ts",
