+++ lib/utils/cache/package.json
@@ -23,14 +23,14 @@
   "repository": "github:exteranto/cache",
   "license": "MIT",
   "dependencies": {
-    "@exteranto/ioc": "^1.1.2",
+    "@exteranto/core": "^1.1.2",
     "@exteranto/storage": "^1.1.2",
     "md5-typescript": "^1.0.5",
     "typescript": "^2.9.2"
   },
   "devDependencies": {
     "@exteranto/core": "^1.1.2",
-    "@exteranto/support": "^1.1.2",
+    "@exteranto/core": "^1.1.2",
     "@types/chai": "^4.1.7",
     "@types/chai-as-promised": "^7.1.0",
     "@types/mocha": "^5.2.5",
