+++ lib/utils/compatibility/package.json
@@ -24,8 +24,8 @@
   "license": "MIT",
   "dependencies": {
     "@exteranto/exceptions": "^1.1.2",
-    "@exteranto/ioc": "^1.1.2",
-    "@exteranto/support": "^1.1.2",
+    "@exteranto/core": "^1.1.2",
+    "@exteranto/core": "^1.1.2",
     "typescript": "^2.9.2"
   },
   "devDependencies": {
