+++ lib/utils/compatibility/src/VersionResolver.ts
@@ -1,4 +1,4 @@
-import { Singleton } from '@exteranto/ioc'
+import { Singleton } from '@exteranto/core'
 
 @Singleton
 export class VersionResolver {
