+++ lib/utils/compatibility/src/Versioning.ts
@@ -1,5 +1,5 @@
 import { VersionNotMatchedException } from '@exteranto/exceptions'
-import { Autowired, Binding, Param } from '@exteranto/ioc'
+import { Autowired, Binding, Param } from '@exteranto/core'
 import { VersionResolver } from './VersionResolver'
 
 @Binding
